Release/15.0

+++ files/dot.profile
@@ -15,7 +15,9 @@ export ENV=~/.shrc
 
 ls -t ~/.serverauth.* 2>/dev/null | tail -n +2 | xargs rm -f --
 
-kbdcontrol -f 1 'startx'
-kbdcontrol -f 3 'lock -np'
-kbdcontrol -f 4 '/etc/rc.suspend'
-kbdcontrol -f 12 'shutdown -p now'
+if [ -z "$SSH_CONNECTION" ]; then
+	kbdcontrol -f 1 'startx'
+	kbdcontrol -f 3 'lock -np'
+	kbdcontrol -f 4 '/etc/rc.suspend'
+	kbdcontrol -f 12 'shutdown -p now'
+fi
